backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "cassandra-driver>=3.29.3",
    "celery>=5.6.3",
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "django-ratelimit>=4.1.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.3.0",
    "kafka-python>=2.3.0",
    "psycopg[binary]>=3.3.3",
    "redis>=7.4.0",
]

[dependency-groups]
dev = [
    "coverage>=7.13.5",
    "django-stubs>=6.0.2",
    "djangorestframework-stubs>=3.16.9",
    "mypy>=1.20.0",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "ruff>=0.15.9",
    "types-redis>=4.6.0.20241004",
]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["test_*.py", "tests.py"]
